+++ CHANGES.rst
@@ -12,6 +12,10 @@ Pre-release (develop branch)
 * Fix `#13 <https://github.com/jantman/awslimitchecker/issues/13>`_ re-enable Sphinx documentation linkcheck
 * Fix `#40 <https://github.com/jantman/awslimitchecker/issues/40>`_ add support for pagination of API responses (to get all results) and handle pagination for all current services
 * Fix `#88 <https://github.com/jantman/awslimitchecker/issues/88>`_ add support for API-derived limits. This is a change to the public API for ``awslimitchecker.limit.AwsLimit`` and the CLI output.
+* Fix `#72 <https://github.com/jantman/awslimitchecker/issues/72>`_ add support for some new limits returned by Trusted Advisor. This renames the following limits:
+  * ``EC2/EC2-VPC Elastic IPs`` to ``EC2/VPC Elastic IP addresses (EIPs)``
+  * ``RDS/Read Replicas per Master`` to ``RDS/Read replicas per master``
+  * ``RDS/Parameter Groups`` to ``RDS/DB parameter groups``
 
 0.1.3 (2015-10-04)
 ------------------